+++ CPSettingsLocations/views.py
@@ -7,11 +7,16 @@ from django.http import HttpResponseRedirect
 
 @login_required(login_url='/account/login')
 def location_settings_view(request):
-    context= None
-    return render(request, 'CPSettingsLocations/view.html', context)
+    if request.user.profile.restaurante is None:
+        messages.success(request, "Antes de editar tus parámetros debes crear tu empresa.")
+        return redirect('settings_rest_create')
+    return render(request, 'CPSettingsLocations/view.html')
 
 @login_required(login_url='/account/login')
 def location_settings_create(request):
+    if request.user.profile.restaurante is None:
+        messages.success(request, "Antes de editar tus parámetros debes crear tu empresa.")
+        return redirect('settings_rest_create')
     if(request.POST):
         if(request.user.profile.restaurante.direcciones.all().count() < request.user.profile.plan.max_empresas or request.user.profile.plan.max_empresas is 0):
             dir = Direccion.objects.create(direccion=request.POST['dir'], lat=request.POST['lat'], long=request.POST['long'], restaurante_id=request.user.profile.restaurante.id)
@@ -26,6 +31,9 @@ def location_settings_create(request):
 
 @login_required(login_url='/account/login')
 def location_settings_delete(request, location_id):
+    if request.user.profile.restaurante is None:
+        messages.success(request, "Antes de editar tus parámetros debes crear tu empresa.")
+        return redirect('settings_rest_create')
     dir = Direccion.objects.get(id=location_id)
     dir.delete()
     messages.success(request, "Dirección eliminada.")
